comments & improvements
